Evaluate TurboQuant KV-cache compression, fix kv_bits plumbing
Fixed _MlxBackend bug: kv_bits was passed to mlx_vlm.load() (silently
ignored) instead of generate() where it actually enables KV-cache
quantization. Also must use kv_quant_scheme="turboquant" — the default
"uniform" scheme crashes on Gemma 4's RotatingKVCache.

Benchmarked kv_bits=4 and kv_bits=8 across 3 transcripts (2K/8K/19K
words). Result: negligible memory savings (<0.1 GB on ~18 GB footprint),
kv_bits=8 is 14% slower. Not worth enabling — leave at None.

Enhanced bench_summarizer.py: multi-transcript support, multiple
--kv-bits values, GPU memory tracking, ROUGE-L quality comparison,
markdown report generation.

+++ scarecrow/config.py
@@ -81,7 +81,7 @@ class Config:
     SUMMARIZER_OUTPUT_BUDGET: int = 4096  # max tokens for summary output
     SUMMARIZER_CHARS_PER_TOKEN: int = 4  # rough heuristic for token estimation
     SUMMARIZER_MLX_MODEL_ID: str = "mlx-community/gemma-4-26b-a4b-it-4bit"
-    SUMMARIZER_MLX_KV_BITS: int | None = None  # TurboQuant: 3-4 for compressed KV cache
+    SUMMARIZER_MLX_KV_BITS: float | None = None  # TurboQuant KV cache quantization bits
 
 
 # Module-level instance — all existing ``config.FOO`` references resolve here.
